[ALPS04715419] Platform: 64bit project build error fix
Kernel fix various build errors for 64 bit project

MTK-Commit-Id: d52f9a4d5d802df3e7ecd4d5fcd164a285fdaf99

CR-Id: ALPS04715419
Feature: Configuration

+++ drivers/misc/mediatek/video/mt6757/dispsys/ddp_ovl.c
@@ -135,7 +135,7 @@ enum CMDQ_EVENT_ENUM ovl_to_cmdq_event_nonsec_end(enum DISP_MODULE_ENUM module)
 		DDPERR("invalid ovl module=%d, get cmdq event nonsecure fail\n",
 		       module);
 		ASSERT(0);
-		return DISP_MODULE_UNKNOWN;
+		return (unsigned long long int)DISP_MODULE_UNKNOWN;
 	}
 	return 0;
 }
